+++ tubular/capping.py
@@ -286,14 +286,12 @@ class BaseCappingTransformer(BaseNumericTransformer, WeightColumnMixin):
 
         X = _convert_dataframe_to_narwhals(X)
 
-        backend = nw.get_native_namespace(X)
-
         weights_column = self.weights_column
         if self.weights_column is None:
             X, weights_column = WeightColumnMixin._create_unit_weights_column(
                 X,
-                backend=backend.__name__,
                 return_native=False,
+                verbose=self.verbose,
             )
         WeightColumnMixin.check_weights_column(self, X, weights_column)
 
@@ -1017,15 +1015,13 @@ class OutOfRangeNullTransformer(BaseCappingTransformer):
 
         super().fit(X=X, y=y)
 
-        backend = nw.get_native_namespace(X)
-
         original_weights_column = self.weights_column
         weights_column = original_weights_column
         if self.weights_column is None:
             X, weights_column = WeightColumnMixin._create_unit_weights_column(
                 X,
-                backend=backend.__name__,
                 return_native=False,
+                verbose=self.verbose,
             )
         WeightColumnMixin.check_weights_column(self, X, weights_column)
         valid_weights_filter_expr = WeightColumnMixin.get_valid_weights_filter_expr(