+++ CHANGELOG.rst
@@ -32,7 +32,8 @@ Each individual change should have a link to the pull request after the descript
 
 Changed
 ^^^^^^^
-- placeholder
+- feat: made GroupRareLevelsTransformer lazyframe compatible `#531 <https://github.com/azukds/tubular/issues/531>`_
+- feat: edited dummy unit weight handling to be more lazy friendly
 
 3.0.0 (19/03/2026)
 ------------------
@@ -41,8 +42,6 @@ Changed
 ^^^^^^^
 - feat: made OneHotEncoder lazyframe compatible `#532 <https://github.com/azukds/tubular/issues/532>`_
 - chore: small refactor of BaseCappingTransformer to remove prepare_quantiles method for readability
-- feat: made GroupRareLevelsTransformer lazyframe compatible `#531 <https://github.com/azukds/tubular/issues/531>`_
-- feat: edited dummy unit weight handling to be more lazy friendly
 - chore: bump minimum python to 310, upgrade black for security fix (need for py310 forced a major version upgrade)
 
 2.8.0 (23/02/2026)