+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changed
 - feat: made GroupRareLevelsTransformer lazyframe compatible `#531 <https://github.com/azukds/tubular/issues/531>`_
 - feat: edited dummy unit weight handling to be more lazy friendly
 - feat: updated ModeImputer to work with lazyframes `#530 <https://github.com/azukds/tubular/issues/530>_`
+- feat: converted MeanImputer to work with lazyframes `#530 <https://github.com/azukds/tubular/issues/530>_`
 
 3.0.0 (19/03/2026)
 ------------------
@@ -44,7 +45,6 @@ Changed
 - feat: made OneHotEncoder lazyframe compatible `#532 <https://github.com/azukds/tubular/issues/532>`_
 - chore: small refactor of BaseCappingTransformer to remove prepare_quantiles method for readability
 - chore: bump minimum python to 310, upgrade black for security fix (need for py310 forced a major version upgrade)
-- feat: converted MeanImputer to work with lazyframes `#530 <https://github.com/azukds/tubular/issues/530>_`
 
 2.8.0 (23/02/2026)
 ------------------