+++ CHANGELOG.rst
@@ -32,7 +32,10 @@ Each individual change should have a link to the pull request after the descript
 
 Changed
 ^^^^^^^
-- placeholder
+- feat: made GroupRareLevelsTransformer lazyframe compatible `#531 <https://github.com/azukds/tubular/issues/531>`_
+- feat: edited dummy unit weight handling to be more lazy friendly
+- feat: updated ModeImputer to work with lazyframes `#530 <https://github.com/azukds/tubular/issues/530>_`
+- feat: converted MeanImputer to work with lazyframes `#530 <https://github.com/azukds/tubular/issues/530>_`
 
 3.0.0 (19/03/2026)
 ------------------
@@ -78,6 +81,13 @@ Changed
 - bugfix - tests with staticmethod decorator were not being detected for earlier python versions, have removed this from test files
 - feat: add module for serializing and deserializing pipelines `#588 <https://github.com/azukds/tubular/issues/588>_`
 - bug: fix issue with rare_levels_record_ attr in to_json method of GroupRareLevelsTransformer
+<<<<<<< HEAD
+<<<<<<< HEAD
+>>>>>>> feature/lazy_weightsmixin
+=======
+>>>>>>> main
+=======
+>>>>>>> feature/lazy_grouprare
 
 2.5.0 (16/12/2025)
 ------------------